requirements.txt
allennlp-models
python==3.11.0
numpy==1.24.1
pandas==1.5.3
bs4==0.0.1
pathlib==1.0.1
requests==2.28.2
transformers
backoff==2.2.1
sqlalchemy==2.0.2
pytest==7.2.1
mysqlclient==2.1.1
mysql-connector-python==8.0.32
torchmetrics==0.11.1
dict_hash==1.1.29
scikit-learn==1.2.2
nltk==3.8.1
python-dotenv==1.0.0
tensorboard==2.14.0
ipywidgets==8.1.0
accelerate==0.21.0
peft==0.4.0
bitsandbytes
trl==0.4.7
sentencepiece==0.1.99
bitsandbytes-cuda116
protobuf
spacy
openai